Using comptime instead of unplugin-macros
