+++ apps/typegpu-docs/astro.config.mjs
@@ -8,7 +8,7 @@ import { defineConfig } from 'astro/config';
 import starlightBlog from 'starlight-blog';
 import starlightTypeDoc, { typeDocSidebarGroup } from 'starlight-typedoc';
 import typegpu from 'unplugin-typegpu/rollup';
-import macros from 'unplugin-macros/vite';
+import { comptime } from 'comptime/vite';
 import { imagetools } from 'vite-imagetools';
 import wasm from 'vite-plugin-wasm';
 import basicSsl from '@vitejs/plugin-basic-ssl';
@@ -55,7 +55,10 @@ export default defineConfig({
       tailwindVite(),
       typegpu({ include: [/\.m?[jt]sx?/] }),
       imagetools(),
-      macros(),
+      {
+        ...comptime(),
+        enforce: 'post',
+      },
       {
         ...basicSsl(),
         apply(_, { mode }) {
