+++ apps/typegpu-docs/package.json
@@ -84,11 +84,11 @@
     "@webgpu/types": "catalog:types",
     "astro-vtbot": "^2.1.10",
     "autoprefixer": "^10.4.21",
+    "comptime": "^0.1.0",
     "magic-string": "^0.30.21",
     "tailwindcss": "^4.1.11",
     "tailwindcss-motion": "^1.1.1",
     "typegpu-testing-utility": "workspace:*",
-    "unplugin-macros": "^0.20.0",
     "vite-imagetools": "catalog:frontend",
     "vite-plugin-wasm": "^3.5.0",
     "yaml": "^2.8.3"
