+++ apps/typegpu-docs/src/components/stackblitz/pnpmWorkspace.ts
@@ -1,30 +1,36 @@
 import { type } from 'arktype';
+import { comptime } from 'comptime';
 import { parse } from 'yaml';
-import pnpmWorkspace from '../../../../../pnpm-workspace.yaml?raw';
 
-const pnpmWorkspaceYamlResult = type({
-  catalogs: {
-    types: {
-      typescript: 'string',
-      '@webgpu/types': 'string',
-      '@types/three': 'string',
-    },
-    test: {
-      vitest: 'string',
-    },
-    frontend: {
-      'vite-imagetools': 'string',
-      'fuse.js': 'string',
-    },
-    example: {
-      'wgpu-matrix': 'string',
-      three: 'string',
-    },
-  },
-})(parse(pnpmWorkspace));
+export const pnpmWorkspaceYaml = comptime(async () => {
+  const pnpmWorkspaceContent = (await import('../../../../../pnpm-workspace.yaml?raw')).default;
 
-if (pnpmWorkspaceYamlResult instanceof type.errors) {
-  throw new Error(pnpmWorkspaceYamlResult.summary);
-}
+  const pnpmWorkspaceYamlType = type({
+    catalogs: {
+      types: {
+        typescript: 'string',
+        '@webgpu/types': 'string',
+        '@types/three': 'string',
+      },
+      test: {
+        vitest: 'string',
+      },
+      frontend: {
+        'vite-imagetools': 'string',
+        'fuse.js': 'string',
+      },
+      example: {
+        'wgpu-matrix': 'string',
+        three: 'string',
+      },
+    },
+  });
 
-export const pnpmWorkspaceYaml = pnpmWorkspaceYamlResult;
+  const pnpmWorkspaceYamlResult = pnpmWorkspaceYamlType(parse(pnpmWorkspaceContent));
+
+  if (pnpmWorkspaceYamlResult instanceof type.errors) {
+    throw new Error(pnpmWorkspaceYamlResult.summary);
+  }
+
+  return pnpmWorkspaceYamlResult;
+});
