+++ apps/typegpu-docs/src/examples/exampleContent.ts
@@ -11,14 +11,13 @@ import type {
   PendingExampleSrcFile,
   ThumbnailPair,
 } from '../utils/examples/types.ts';
-import { pathToExampleKey } from './macros.ts';
+import { pathToExampleKey } from './pathToExampleKey.ts';
 import { metaContent, tsContentLazy, tsnotoverContentLazy } from './importers.ts';
 
 // ----
 // Precomputing APIs used across examples. This is done at build-time to avoid loading
 // all example source code into the browser unless its required.
-import { MACRO_computeUsedApis } from './macros.ts' with { type: 'macro' };
-const usedApis = await MACRO_computeUsedApis();
+import { usedApis } from './usedApis.ts';
 // ----
 
 function extractUrlFromViteImport(importFn: () => void): [URL | undefined, boolean] {
